+++ src/us/kbase/workspace/test/controllers/handle/HandleServiceController.java
@@ -107,7 +107,7 @@ public class HandleServiceController {
 		env.put("KB_DEPLOYMENT_CONFIG", hmIniFile.toString());
 		handleManager = handlemgrpb.start();
 		
-		Thread.sleep(2000); //let the manager start up
+		Thread.sleep(15000); //let the manager start up
 		
 	}