+++ docsource/releasenotes.rst
@@ -4,7 +4,14 @@ Workspace service release notes
 VERSION: 0.7.1 (Released TBD)
 --------------------------------
 
+BACKWARDS INCOMPATIBILITIES:
+
+* The ``getPermissions`` administration command, like the ``get_permissions`` method, is now
+  deprecated.
+
 NEW FEATURES:
+
+* Added the ``getPermissionsMass`` administration command.
 
 UPDATED FEATURES / MAJOR BUG FIXES: